refactor cook web operator user detail page
