+++ docs/exec-plans/index.md
@@ -16,6 +16,7 @@ Active and completed ExecPlans live here. The structure and required sections ar
 - [Observability Artifacts, Consistency Probes, and Frontend Trace IDs](./active/observability-artifacts-consistency-frontend-trace.md)
 - [Operator Course Detail Tab Splitting](./active/operator-course-detail-tab-splitting.md)
 - [ExecPlan: Operator Credit Grant Package](./active/operator-credit-grant-package.md)
+- [Operator User Detail Page Slimming](./active/operator-user-detail-page-slimming.md)
 
 ## Completed
