+++ src/cook-web/src/app/admin/operations/users/[user_bid]/page.tsx
@@ -1,104 +1,32 @@
 'use client';
 
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useCallback, useEffect, useRef, useState } from 'react';
 import { useParams, useRouter } from 'next/navigation';
-import { CircleHelp } from 'lucide-react';
 import { useTranslation } from 'react-i18next';
 import api from '@/api';
 import AdminTitle from '@/app/admin/components/AdminTitle';
-import { formatAdminCredits } from '@/app/admin/lib/numberFormat';
-import { useEnvStore } from '@/c-store';
-import type { EnvStoreState } from '@/c-types/store';
 import ErrorDisplay from '@/components/ErrorDisplay';
 import Loading from '@/components/loading';
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from '@/components/ui/tooltip';
-import { resolveContactMode } from '@/lib/resolve-contact-mode';
-import { ErrorWithCode } from '@/lib/request';
+import { TooltipProvider } from '@/components/ui/tooltip';
 import AdminOperationsBreadcrumb from '../../AdminOperationsBreadcrumb';
-import { formatOperatorUtcDateTime } from '../dateTime';
-import type {
-  AdminOperationUserCourseItem,
-  AdminOperationUserCreditFilters,
-  AdminOperationUserCreditSummary,
-  AdminOperationUserCreditUsageDetailResponse,
-  AdminOperationUserCreditsResponse,
-  AdminOperationUserDetailResponse,
-} from '../../operation-user-types';
+import type { AdminOperationUserCreditUsageDetailResponse } from '../../operation-user-types';
 import useOperatorGuard from '../../useOperatorGuard';
 import UserDetailSummarySection from './UserDetailSummarySection';
 import UserDetailTabsSection from './UserDetailTabsSection';
+import useUserCreditLedgerData from './useUserCreditLedgerData';
+import useUserDetailData from './useUserDetailData';
+import useUserDetailViewModel, {
+  formatLearningProgress,
+} from './useUserDetailViewModel';
 import {
-  createUserCreditFilters,
-  FILTER_ALL_OPTION,
-  sanitizeCreditFiltersByType,
-} from './creditFilterUtils';
+  DETAIL_TAB_HASHES,
+  EMPTY_VALUE,
+  resolveDetailTabFromHash,
+  type DetailTab,
+} from './userDetailConstants';
 
-type ErrorState = { message: string; code?: number };
-type DetailTab = 'credits' | 'learning' | 'created';
-
-const CREDITS_PAGE_SIZE = 10;
-const EMPTY_VALUE = '--';
-const DETAIL_TAB_HASHES: Record<DetailTab, string> = {
-  credits: '#credits',
-  learning: '#learning-courses',
-  created: '#created-courses',
-};
-const resolveDetailTabFromHash = (hash: string): DetailTab | null => {
-  const hashEntry = Object.entries(DETAIL_TAB_HASHES).find(
-    ([, targetHash]) => targetHash === hash,
-  ) as [DetailTab, string] | undefined;
-
-  return hashEntry?.[0] ?? null;
-};
-const resolveCourseCount = (
-  count: number,
-  courses?: AdminOperationUserCourseItem[],
-) => (count > 0 ? count : (courses || []).length);
-const DEFAULT_CREDIT_SUMMARY: AdminOperationUserCreditSummary = {
-  available_credits: '',
-  subscription_credits: '',
-  topup_credits: '',
-  credits_expire_at: '',
-  has_active_subscription: false,
-};
-const createEmptyCreditsResponse = (): AdminOperationUserCreditsResponse => ({
-  summary: DEFAULT_CREDIT_SUMMARY,
-  items: [],
-  page: 1,
-  page_count: 0,
-  page_size: CREDITS_PAGE_SIZE,
-  total: 0,
-});
-const EMPTY_DETAIL: AdminOperationUserDetailResponse = {
-  user_bid: '',
-  mobile: '',
-  email: '',
-  nickname: '',
-  user_status: 'unknown',
-  user_role: 'unknown',
-  user_roles: [],
-  login_methods: [],
-  language: '',
-  learning_courses: [],
-  learning_course_count: 0,
-  created_courses: [],
-  created_course_count: 0,
-  registration_source: 'unknown',
-  total_paid_amount: '0',
-  available_credits: '',
-  subscription_credits: '',
-  topup_credits: '',
-  credits_expire_at: '',
-  has_active_subscription: false,
-  last_login_at: '',
-  last_learning_at: '',
-  created_at: '',
-  updated_at: '',
+type UserBidParams = {
+  user_bid: string;
 };
 
 /**
@@ -211,138 +139,44 @@ const EMPTY_DETAIL: AdminOperationUserDetailResponse = {
  * t('module.user.defaultUserName')
  */
 
-const formatLearningProgress = (
-  course: AdminOperationUserCourseItem,
-): string => {
-  const totalLessonCount = Number(course.total_lesson_count || 0);
-  if (!Number.isFinite(totalLessonCount) || totalLessonCount <= 0) {
-    return EMPTY_VALUE;
-  }
-
-  const completedLessonCount = Math.max(
-    0,
-    Math.min(Number(course.completed_lesson_count || 0), totalLessonCount),
-  );
-  const progressPercent = Math.round(
-    (completedLessonCount / totalLessonCount) * 100,
-  );
-  return `${progressPercent}% (${completedLessonCount}/${totalLessonCount})`;
-};
-
 export default function AdminOperationUserDetailPage() {
-  const { t, i18n } = useTranslation();
   const { t: tOperationsUsers } = useTranslation('module.operationsUser');
-  const { t: tOperationsCourse } = useTranslation('module.operationsCourse');
   const router = useRouter();
-  const params = useParams<{ user_bid: string }>();
+  const params = useParams<UserBidParams>();
   const { isReady } = useOperatorGuard();
-  const loginMethodsEnabled = useEnvStore(
-    (state: EnvStoreState) => state.loginMethodsEnabled,
-  );
-  const defaultLoginMethod = useEnvStore(
-    (state: EnvStoreState) => state.defaultLoginMethod,
-  );
-  const currencySymbol = useEnvStore(
-    (state: EnvStoreState) => state.currencySymbol || '',
-  );
-  const defaultUserName = useMemo(() => t('module.user.defaultUserName'), [t]);
   const detailTabsSectionRef = useRef<HTMLDivElement | null>(null);
-  const hasInitializedCreditStateRef = useRef(false);
-  const [detailLoading, setDetailLoading] = useState(true);
-  const [detailError, setDetailError] = useState<ErrorState | null>(null);
-  const [creditsLoading, setCreditsLoading] = useState(true);
-  const [creditsError, setCreditsError] = useState<ErrorState | null>(null);
-  const [detailRetryNonce, setDetailRetryNonce] = useState(0);
-  const [creditsRetryNonce, setCreditsRetryNonce] = useState(0);
-  const [creditsPageIndex, setCreditsPageIndex] = useState(1);
-  const [creditFiltersDraft, setCreditFiltersDraft] =
-    useState<AdminOperationUserCreditFilters>(createUserCreditFilters);
-  const [creditFilters, setCreditFilters] =
-    useState<AdminOperationUserCreditFilters>(createUserCreditFilters);
+  const hasInitializedDetailTabRef = useRef(false);
   const [activeTab, setActiveTab] = useState<DetailTab>('credits');
-  const [detail, setDetail] =
-    useState<AdminOperationUserDetailResponse>(EMPTY_DETAIL);
-  const [credits, setCredits] = useState<AdminOperationUserCreditsResponse>(
-    createEmptyCreditsResponse,
-  );
+  const {
+    detail,
+    detailLoading,
+    detailError,
+    retryDetail,
+    userBid,
+    userBidErrorMessage,
+  } = useUserDetailData({
+    isReady,
+    rawUserBid: params?.user_bid,
+  });
+  const {
+    credits,
+    creditsLoading,
+    creditsError,
+    creditsPageIndex,
+    creditFilters,
+    creditFiltersDraft,
+    setCreditFiltersDraft,
+    setCreditsPageIndex,
+    handleCreditSearch,
+    handleCreditTypeChange,
+    handleCreditReset,
+    retryCredits,
+  } = useUserCreditLedgerData({
+    isReady,
+    userBid,
+    userBidErrorMessage,
+  });
 
-  const userBidState = useMemo(() => {
-    const rawUserBid = String(params?.user_bid || '').trim();
-    if (!rawUserBid) {
-      return {
-        userBid: '',
-        errorMessage: t('server.common.paramsError'),
-      };
-    }
-
-    try {
-      return {
-        userBid: decodeURIComponent(rawUserBid),
-        errorMessage: '',
-      };
-    } catch {
-      return {
-        userBid: '',
-        errorMessage: t('server.common.paramsError'),
-      };
-    }
-  }, [params, t]);
-  const userBid = userBidState.userBid;
-  const contactType = useMemo(
-    () => resolveContactMode(loginMethodsEnabled, defaultLoginMethod),
-    [defaultLoginMethod, loginMethodsEnabled],
-  );
-  const contactLabel = useMemo(
-    () =>
-      contactType === 'email'
-        ? tOperationsUsers('table.email')
-        : tOperationsUsers('table.mobile'),
-    [contactType, tOperationsUsers],
-  );
-  const contactValue = useMemo(
-    () =>
-      contactType === 'email'
-        ? detail.email || detail.mobile
-        : detail.mobile || detail.email,
-    [contactType, detail.email, detail.mobile],
-  );
-  const creditOwnerLabel = useMemo(() => {
-    const mobile = detail.mobile || '';
-    const displayName = detail.nickname || contactValue || userBid || '';
-    if (mobile && displayName && displayName !== mobile) {
-      return `${mobile} / ${displayName}`;
-    }
-    return mobile || displayName || EMPTY_VALUE;
-  }, [contactValue, detail.mobile, detail.nickname, userBid]);
-  const creditSummary = useMemo<AdminOperationUserCreditSummary>(
-    () => ({
-      available_credits:
-        credits.summary.available_credits || detail.available_credits || '',
-      subscription_credits:
-        credits.summary.subscription_credits ||
-        detail.subscription_credits ||
-        '',
-      topup_credits:
-        credits.summary.topup_credits || detail.topup_credits || '',
-      credits_expire_at:
-        credits.summary.credits_expire_at || detail.credits_expire_at || '',
-      has_active_subscription:
-        credits.summary.has_active_subscription ||
-        detail.has_active_subscription,
-    }),
-    [
-      credits.summary.available_credits,
-      credits.summary.credits_expire_at,
-      credits.summary.has_active_subscription,
-      credits.summary.subscription_credits,
-      credits.summary.topup_credits,
-      detail.available_credits,
-      detail.credits_expire_at,
-      detail.has_active_subscription,
-      detail.subscription_credits,
-      detail.topup_credits,
-    ],
-  );
   const scrollToDetailTabsSection = useCallback(() => {
     detailTabsSectionRef.current?.scrollIntoView({
       behavior: 'smooth',
@@ -370,186 +204,25 @@ export default function AdminOperationUserDetailPage() {
     },
     [scrollToDetailTabsSection, syncDetailTabHash],
   );
-
-  useEffect(() => {
-    if (!isReady) {
-      return;
-    }
-
-    if (userBidState.errorMessage) {
-      setDetailError({ message: userBidState.errorMessage });
-      setDetailLoading(false);
-      return;
-    }
-
-    let cancelled = false;
-
-    const fetchDetail = async () => {
-      setDetailLoading(true);
-      setDetailError(null);
-      try {
-        const response = (await api.getAdminOperationUserDetail({
-          user_bid: userBid,
-        })) as AdminOperationUserDetailResponse;
-        if (cancelled) {
-          return;
-        }
-        setDetail(response);
-      } catch (requestError) {
-        if (cancelled) {
-          return;
-        }
-        const resolvedError = requestError as ErrorWithCode;
-        setDetailError({
-          message: resolvedError.message || t('common.core.networkError'),
-          code: resolvedError.code,
-        });
-      } finally {
-        if (!cancelled) {
-          setDetailLoading(false);
-        }
-      }
-    };
-
-    void fetchDetail();
-
-    return () => {
-      cancelled = true;
-    };
-  }, [detailRetryNonce, isReady, t, userBid, userBidState.errorMessage]);
-
-  useEffect(() => {
-    if (!hasInitializedCreditStateRef.current) {
-      hasInitializedCreditStateRef.current = true;
-      return;
-    }
-    setCreditsPageIndex(1);
-    setCreditsError(null);
-    setCredits(createEmptyCreditsResponse());
-    setCreditFiltersDraft(createUserCreditFilters());
-    setCreditFilters(createUserCreditFilters());
-    setActiveTab('credits');
-    syncDetailTabHash('credits');
-  }, [syncDetailTabHash, userBid]);
-
-  useEffect(() => {
-    if (!isReady || !userBid || userBidState.errorMessage) {
-      return;
-    }
-
-    let cancelled = false;
-
-    const fetchCredits = async () => {
-      setCreditsLoading(true);
-      setCreditsError(null);
-      try {
-        const response = (await api.getAdminOperationUserCredits({
-          user_bid: userBid,
-          page_index: creditsPageIndex,
-          page_size: CREDITS_PAGE_SIZE,
-          credit_type:
-            creditFilters.creditType === FILTER_ALL_OPTION
-              ? ''
-              : creditFilters.creditType,
-          grant_source:
-            creditFilters.creditType === 'grant' &&
-            creditFilters.grantSource !== FILTER_ALL_OPTION
-              ? creditFilters.grantSource
-              : '',
-          course_query:
-            creditFilters.creditType === 'consume'
-              ? creditFilters.courseQuery.trim()
-              : '',
-          usage_scene:
-            creditFilters.creditType === 'consume' &&
-            creditFilters.usageScene !== FILTER_ALL_OPTION
-              ? creditFilters.usageScene
-              : '',
-          usage_mode:
-            creditFilters.creditType === 'consume' &&
-            creditFilters.usageMode !== FILTER_ALL_OPTION
-              ? creditFilters.usageMode
-              : '',
-          start_time:
-            creditFilters.creditType !== FILTER_ALL_OPTION
-              ? creditFilters.startTime
-              : '',
-          end_time:
-            creditFilters.creditType !== FILTER_ALL_OPTION
-              ? creditFilters.endTime
-              : '',
-        })) as AdminOperationUserCreditsResponse;
-        if (cancelled) {
-          return;
-        }
-        setCredits(response);
-      } catch (requestError) {
-        if (cancelled) {
-          return;
-        }
-        const resolvedError = requestError as ErrorWithCode;
-        setCreditsError({
-          message: resolvedError.message || t('common.core.networkError'),
-          code: resolvedError.code,
-        });
-        setCredits(current => ({
-          ...current,
-          items: [],
-          page: creditsPageIndex,
-          page_count: 0,
-          total: 0,
-        }));
-      } finally {
-        if (!cancelled) {
-          setCreditsLoading(false);
-        }
-      }
-    };
-
-    void fetchCredits();
-
-    return () => {
-      cancelled = true;
-    };
-  }, [
-    creditsPageIndex,
-    creditFilters,
-    creditsRetryNonce,
-    isReady,
-    t,
+  const handleLearningCoursesClick = useCallback(() => {
+    setDetailTab('learning', { scrollToSection: true });
+  }, [setDetailTab]);
+  const handleCreatedCoursesClick = useCallback(() => {
+    setDetailTab('created', { scrollToSection: true });
+  }, [setDetailTab]);
+  const {
+    basicInfoItems,
+    overviewItems,
+    creditsOverviewItems,
+    creditOwnerLabel,
+    resolveCourseStatusLabel,
+  } = useUserDetailViewModel({
+    detail,
+    credits,
     userBid,
-    userBidState.errorMessage,
-  ]);
-
-  const handleCreditSearch = () => {
-    const nextFilters = sanitizeCreditFiltersByType({
-      ...creditFiltersDraft,
-      courseQuery: creditFiltersDraft.courseQuery.trim(),
-    });
-    setCreditFiltersDraft(nextFilters);
-    setCreditFilters(nextFilters);
-    setCreditsPageIndex(1);
-  };
-
-  const handleCreditTypeChange = useCallback(
-    (nextFilters: AdminOperationUserCreditFilters) => {
-      const sanitizedFilters = sanitizeCreditFiltersByType({
-        ...nextFilters,
-        courseQuery: nextFilters.courseQuery.trim(),
-      });
-      setCreditFiltersDraft(sanitizedFilters);
-      setCreditFilters(sanitizedFilters);
-      setCreditsPageIndex(1);
-    },
-    [],
-  );
-
-  const handleCreditReset = () => {
-    const nextFilters = createUserCreditFilters();
-    setCreditFiltersDraft(nextFilters);
-    setCreditFilters(nextFilters);
-    setCreditsPageIndex(1);
-  };
+    onLearningCoursesClick: handleLearningCoursesClick,
+    onCreatedCoursesClick: handleCreatedCoursesClick,
+  });
 
   const handleCreditCourseOpen = useCallback(
     (courseBid: string) => {
@@ -567,6 +240,15 @@ export default function AdminOperationUserDetailPage() {
   );
 
   useEffect(() => {
+    if (!hasInitializedDetailTabRef.current) {
+      hasInitializedDetailTabRef.current = true;
+      return;
+    }
+    setActiveTab('credits');
+    syncDetailTabHash('credits');
+  }, [syncDetailTabHash, userBid]);
+
+  useEffect(() => {
     if (typeof window === 'undefined' || detailLoading) {
       return;
     }
@@ -580,218 +262,6 @@ export default function AdminOperationUserDetailPage() {
     scrollToDetailTabsSection();
   }, [detailLoading, scrollToDetailTabsSection]);
 
-  const resolveRoleLabel = useCallback(
-    (role: string) => tOperationsUsers(`roleLabels.${role || 'unknown'}`),
-    [tOperationsUsers],
-  );
-  const resolveRegistrationSourceLabel = useCallback(
-    (source: string) =>
-      tOperationsUsers(`registrationSourceLabels.${source || 'unknown'}`),
-    [tOperationsUsers],
-  );
-  const resolveCourseStatusLabel = (status: string) => {
-    if (status === 'published') {
-      return tOperationsCourse('statusLabels.published');
-    }
-    if (status === 'unpublished') {
-      return tOperationsCourse('statusLabels.unpublished');
-    }
-    const unknownLabel = tOperationsCourse('statusLabels.unknown');
-    return status ? `${unknownLabel} (${status})` : unknownLabel;
-  };
-  const resolveCreditsExpireAt = useCallback(() => {
-    if (creditSummary.credits_expire_at) {
-      return formatOperatorUtcDateTime(creditSummary.credits_expire_at);
-    }
-    if (Number(creditSummary.available_credits || 0) > 0) {
-      return tOperationsUsers('credits.longTerm');
-    }
-    return EMPTY_VALUE;
-  }, [
-    creditSummary.available_credits,
-    creditSummary.credits_expire_at,
-    tOperationsUsers,
-  ]);
-  const creditExpireAtLabel = useMemo(
-    () => (
-      <>
-        <span>
-          {tOperationsUsers('detail.creditsOverviewLabels.creditsExpireAt')}
-        </span>
-        <Tooltip>
-          <TooltipTrigger asChild>
-            <button
-              type='button'
-              aria-label={tOperationsUsers(
-                'detail.creditExpireAtHintAriaLabel',
-              )}
-              className='inline-flex h-4 w-4 items-center justify-center rounded-full text-muted-foreground transition-colors hover:text-foreground'
-            >
-              <CircleHelp className='h-3.5 w-3.5' />
-            </button>
-          </TooltipTrigger>
-          <TooltipContent
-            side='top'
-            className='max-w-[220px] text-center'
-          >
-            {tOperationsUsers('detail.creditExpireAtHint')}
-          </TooltipContent>
-        </Tooltip>
-      </>
-    ),
-    [tOperationsUsers],
-  );
-  const basicInfoItems = useMemo(
-    () => [
-      {
-        key: 'contact',
-        label: contactLabel,
-        value: contactValue,
-      },
-      {
-        key: 'nickname',
-        label: tOperationsUsers('table.nickname'),
-        value: detail.nickname || defaultUserName,
-      },
-      {
-        key: 'role',
-        label: tOperationsUsers('table.role'),
-        value: resolveRoleLabel(detail.user_role),
-      },
-      {
-        key: 'registrationSource',
-        label: tOperationsUsers('table.registrationSource'),
-        value: resolveRegistrationSourceLabel(detail.registration_source),
-      },
-      {
-        key: 'lastLoginAt',
-        label: tOperationsUsers('table.lastLoginAt'),
-        value: formatOperatorUtcDateTime(detail.last_login_at),
-      },
-      {
-        key: 'createdAt',
-        label: tOperationsUsers('table.createdAt'),
-        value: formatOperatorUtcDateTime(detail.created_at),
-      },
-    ],
-    [
-      contactLabel,
-      contactValue,
-      defaultUserName,
-      detail.created_at,
-      detail.last_login_at,
-      detail.nickname,
-      detail.registration_source,
-      detail.user_role,
-      resolveRegistrationSourceLabel,
-      resolveRoleLabel,
-      tOperationsUsers,
-    ],
-  );
-  const overviewItems = useMemo(
-    () => [
-      {
-        key: 'totalPaidAmount',
-        label: tOperationsUsers('table.totalPaidAmount'),
-        value: `${currencySymbol}${detail.total_paid_amount || '0'}`,
-      },
-      {
-        key: 'learningCourses',
-        label: tOperationsUsers('table.learningCourses'),
-        value: String(
-          resolveCourseCount(
-            detail.learning_course_count,
-            detail.learning_courses,
-          ),
-        ),
-        valueClassName: 'text-primary',
-        valueAriaLabel: tOperationsUsers('table.learningCourses'),
-        onClick: () => setDetailTab('learning', { scrollToSection: true }),
-      },
-      {
-        key: 'createdCourses',
-        label: tOperationsUsers('table.createdCourses'),
-        value: String(
-          resolveCourseCount(
-            detail.created_course_count,
-            detail.created_courses,
-          ),
-        ),
-        valueClassName: 'text-primary',
-        valueAriaLabel: tOperationsUsers('table.createdCourses'),
-        onClick: () => setDetailTab('created', { scrollToSection: true }),
-      },
-      {
-        key: 'lastLearningAt',
-        label: tOperationsUsers('table.lastLearningAt'),
-        value: formatOperatorUtcDateTime(detail.last_learning_at),
-      },
-    ],
-    [
-      currencySymbol,
-      detail.created_course_count,
-      detail.created_courses,
-      detail.last_learning_at,
-      detail.learning_course_count,
-      detail.learning_courses,
-      detail.total_paid_amount,
-      setDetailTab,
-      tOperationsUsers,
-    ],
-  );
-  const creditsOverviewItems = useMemo(
-    () => [
-      {
-        key: 'availableCredits',
-        label: tOperationsUsers(
-          'detail.creditsOverviewLabels.availableCredits',
-        ),
-        value: creditSummary.available_credits
-          ? formatAdminCredits(
-              Number(creditSummary.available_credits),
-              i18n.language,
-            )
-          : '',
-      },
-      {
-        key: 'subscriptionCredits',
-        label: tOperationsUsers(
-          'detail.creditsOverviewLabels.subscriptionCredits',
-        ),
-        value: creditSummary.subscription_credits
-          ? formatAdminCredits(
-              Number(creditSummary.subscription_credits),
-              i18n.language,
-            )
-          : '',
-      },
-      {
-        key: 'topupCredits',
-        label: tOperationsUsers('detail.creditsOverviewLabels.topupCredits'),
-        value: creditSummary.topup_credits
-          ? formatAdminCredits(
-              Number(creditSummary.topup_credits),
-              i18n.language,
-            )
-          : '',
-      },
-      {
-        key: 'creditsExpireAt',
-        label: creditExpireAtLabel,
-        value: resolveCreditsExpireAt(),
-      },
-    ],
-    [
-      creditExpireAtLabel,
-      creditSummary.available_credits,
-      creditSummary.subscription_credits,
-      creditSummary.topup_credits,
-      i18n.language,
-      resolveCreditsExpireAt,
-      tOperationsUsers,
-    ],
-  );
-
   if (!isReady || detailLoading) {
     return <Loading />;
   }
@@ -802,7 +272,7 @@ export default function AdminOperationUserDetailPage() {
         <ErrorDisplay
           errorCode={detailError.code || 0}
           errorMessage={detailError.message}
-          onRetry={() => setDetailRetryNonce(value => value + 1)}
+          onRetry={retryDetail}
         />
       </div>
     );
@@ -868,8 +338,8 @@ export default function AdminOperationUserDetailPage() {
                   onTypeChange: handleCreditTypeChange,
                   onSearch: handleCreditSearch,
                   onReset: handleCreditReset,
-                  onPageChange: page => setCreditsPageIndex(page),
-                  onRetry: () => setCreditsRetryNonce(value => value + 1),
+                  onPageChange: setCreditsPageIndex,
+                  onRetry: retryCredits,
                   onCourseOpen: handleCreditCourseOpen,
                   onUsageDetailLoad: usageBid =>
                     api.getAdminOperationUserCreditUsageDetail({
